requirements.txt
Django==2.2.3
django-crispy-forms==1.7.2
pytz==2019.1
sqlparse==0.3.0
Django==2.2.3
django-crispy-forms==1.7.2
django-filter==2.1.0
django-widget-tweaks==1.4.5
pytz==2019.1
sqlparse==0.3.0
